Move mi_disp interaction into a header, add macro
